pyproject.toml
[tool.poetry]
name = "telegram_gpt_bot"
version = "0.1.0"
description = ""
authors = ["zeusina <kachusov_k@outlook.com>"]
readme = "README.md"
packages = [{include = "src/"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
pyyaml = "^6.0"
retrying = "^1.3.4"
aiogram = {version = "^3.0.0b7", allow-prereleases = true}
fake-useragent = "^1.1.3"
requests = "^2.29.0"
tls-client = "^0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
